Calendar terminado

+++ src/hooks/useCalendarStore.js
@@ -1,28 +1,51 @@
 import { useDispatch, useSelector } from 'react-redux'
-import { addNewEvent, updateEvent, setActiveEvent, deleteEvent } from '../store'
+import Swal from 'sweetalert2'
+import { calendarApi } from '../api'
+import { convertEventsToDateEvents } from '../helpers'
+import { addNewEvent, updateEvent, setActiveEvent, deleteEvent, setEvents } from '../store'
 
 export const useCalendarStore = () => {
   const dispatch = useDispatch()
   const { events, activeEvent } = useSelector(state => state.calendar)
+  const { user } = useSelector(state => state.auth)
 
   const onSetActiveEvent = (calendarEvent) => {
     dispatch(setActiveEvent(calendarEvent))
   }
 
+  const startLoadingEvents = async () => {
+    try {
+      const { data } = await calendarApi.get('/event')
+      const events = convertEventsToDateEvents(data.events)
+      dispatch(setEvents(events))
+    } catch (error) {
+      console.log(error)
+    }
+  }
+
   const startSavingEvent = async (calendarEvent) => {
-    // TODO: Llegar al backend
-
-    // Todo bien
-
-    if (calendarEvent._id) {
-      dispatch(updateEvent({ ...calendarEvent }))
-    } else {
-      dispatch(addNewEvent({ ...calendarEvent, _id: new Date().getTime() }))
+    try {
+      if (calendarEvent.id) {
+        const { data } = await calendarApi.put(`/event/${calendarEvent.id}`, calendarEvent)
+        dispatch(updateEvent({ ...calendarEvent, id: data.event.id, user }))
+      } else {
+        const { data } = await calendarApi.post('/event', calendarEvent)
+        dispatch(addNewEvent({ ...calendarEvent, id: data.event.id, user }))
+      }
+    } catch (error) {
+      Swal.fire('Error al guardar', error.response.data?.msg, 'error')
     }
   }
 
   const startDeleteEvent = async () => {
-    dispatch(deleteEvent())
+    try {
+      const { data } = await calendarApi.delete(`/event/${activeEvent.id}`)
+      if (data.ok) {
+        dispatch(deleteEvent())
+      }
+    } catch (error) {
+      Swal.fire('Error al eliminar', error.response.data?.msg, 'error')
+    }
   }
 
   return {
@@ -31,6 +54,7 @@ export const useCalendarStore = () => {
     hasEventSelected: !!activeEvent,
     onSetActiveEvent,
     startSavingEvent,
-    startDeleteEvent
+    startDeleteEvent,
+    startLoadingEvents
   }
 }
